refactor(extensions): 调整扩展功能包结构以优化模块划分
- 将技能相关文件迁移至 extensions.skills 包下
- 将工作空间相关文件迁移至 extensions.workspace 包下
- 更新所有相关导入路径，确保引用正确
- 调整部分代码中 MotionEvent 引用方式，统一导入
- 保持项目包结构清晰，便于后续维护和扩展
