+++ app/src/main/java/me/rerere/rikkahub/di/ViewModelModule.kt
@@ -14,10 +14,10 @@ import me.rerere.rikkahub.ui.pages.stats.StatsVM
 import me.rerere.rikkahub.ui.pages.imggen.ImgGenVM
 import me.rerere.rikkahub.ui.pages.extensions.PromptVM
 import me.rerere.rikkahub.ui.pages.extensions.QuickMessagesVM
-import me.rerere.rikkahub.ui.pages.extensions.SkillDetailVM
-import me.rerere.rikkahub.ui.pages.extensions.SkillsVM
-import me.rerere.rikkahub.ui.pages.extensions.WorkspaceDetailVM
-import me.rerere.rikkahub.ui.pages.extensions.WorkspaceVM
+import me.rerere.rikkahub.ui.pages.extensions.skills.SkillDetailVM
+import me.rerere.rikkahub.ui.pages.extensions.skills.SkillsVM
+import me.rerere.rikkahub.ui.pages.extensions.workspace.WorkspaceDetailVM
+import me.rerere.rikkahub.ui.pages.extensions.workspace.WorkspaceVM
 import me.rerere.rikkahub.ui.pages.setting.SettingVM
 import me.rerere.rikkahub.ui.pages.share.handler.ShareHandlerVM
 import me.rerere.rikkahub.ui.pages.translator.TranslatorVM
