+++ app/src/main/java/me/rerere/rikkahub/ui/pages/extensions/skills/SkillDetailPage.kt
@@ -1,4 +1,4 @@
-package me.rerere.rikkahub.ui.pages.extensions
+package me.rerere.rikkahub.ui.pages.extensions.skills
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.clickable
