+++ app/src/main/java/me/rerere/rikkahub/ui/pages/extensions/skills/SkillDetailVM.kt
@@ -1,4 +1,4 @@
-package me.rerere.rikkahub.ui.pages.extensions
+package me.rerere.rikkahub.ui.pages.extensions.skills
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
