+++ app/src/main/java/me/rerere/rikkahub/ui/pages/extensions/workspace/WorkspacePage.kt
@@ -1,4 +1,4 @@
-package me.rerere.rikkahub.ui.pages.extensions
+package me.rerere.rikkahub.ui.pages.extensions.workspace
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
